Сделал папку с контрактами. Колонки для сортировки теперь Enum
