+++ web/Contracts/SearchParameters/SearchParametersExtensions.cs
@@ -1,20 +1,19 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 
-namespace web.Other.SearchParameters
+namespace web.Contracts.SearchParameters
 {
     public static class SearchParametersExtensions
     {
         public static string ToUrlParameter(this ISearchParameter searchParameter)
         {
-            if(searchParameter is null)
+            if (searchParameter is null)
             {
                 return string.Empty;
             }
 
             var parameters = new List<string>();
             searchParameter.ToParametersList(parameters);
-            if(parameters.Count == 0)
+            if (parameters.Count == 0)
             {
                 return string.Empty;
             }
