+++ web/Contracts/SearchParameters/SortingColumns/CompanySortingColumn.cs
@@ -1,0 +1,10 @@
+﻿namespace web.Contracts.SearchParameters.SortingColumns
+{
+    /// <summary>
+    /// Колонки для сортировки компаний.
+    /// </summary>
+    public enum CompanySortingColumn
+    {
+        Name
+    }
+}
