+++ web/Contracts/SearchParameters/SortingColumns/RoleSortingColumn.cs
@@ -1,0 +1,11 @@
+﻿namespace web.Contracts.SearchParameters.SortingColumns
+{
+    /// <summary>
+    /// Колонки для сортировки ролей.
+    /// </summary>
+    public enum RoleSortingColumn
+    {
+        Type,
+        Name
+    }
+}
