+++ web/Contracts/SearchParameters/SortingColumns/UserSortingColumn.cs
@@ -1,0 +1,12 @@
+﻿namespace web.Contracts.SearchParameters.SortingColumns
+{
+    /// <summary>
+    /// Колонки для сортировки пользователей.
+    /// </summary>
+    public enum UserSortingColumn
+    {
+        FirstName,
+        LastName,
+        Email
+    }
+}
