+++ web/Contracts/SearchParameters/UserSearchParameters.cs
@@ -1,12 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
 using System.Web;
-using Microsoft.EntityFrameworkCore.Storage.ValueConversion.Internal;
+using web.Contracts.SearchParameters.SortingColumns;
 using web.Entities;
 
-namespace web.Other.SearchParameters
+namespace web.Contracts.SearchParameters
 {
     public class UserSearchParameters
         : DefaultSearchParameters
@@ -28,6 +27,17 @@ namespace web.Other.SearchParameters
         /// </summary>
         public RoleType? Role { get; set; }
 
+        /// <summary>
+        /// Имя колонки для сортировки.
+        /// </summary>
+        [MaxLength(32)]
+        public UserSortingColumn? SortingColumn { get; set; }
+
+        /// <summary>
+        /// Направление сортировки.
+        /// </summary>
+        public ListSortDirection? SortDirection { get; set; }
+
         public override void ToParametersList(List<string> parameters)
         {
             base.ToParametersList(parameters);
@@ -43,6 +53,10 @@ namespace web.Other.SearchParameters
             if (this.Role is not null)
             {
                 parameters.Add($"{HttpUtility.UrlEncode(nameof(this.Role))}={HttpUtility.UrlEncode(this.Role.ToString())}");
+            }
+            if (this.SortingColumn is not null)
+            {
+                parameters.Add($"{HttpUtility.UrlEncode(nameof(this.SortingColumn))}={HttpUtility.UrlEncode(this.SortingColumn.ToString())}");
             }
         }
     }
