+++ web/Repositories/IUserRepository.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using web.Dto;
-using web.Entities;
-using web.Other.SearchParameters;
-using web.Dto.Request;
-using web.Dto.Response;
+using web.Contracts.Dto.Request;
+using web.Contracts.Dto.Response;
+using web.Contracts.SearchParameters;
 
 namespace web.Repositories
 {
